Implement waiting for cache lock
Also include some refactoring

+++ vinfo/cache.py
@@ -3,10 +3,22 @@ import itertools
 import logging
 import os
 from pathlib import Path
+import time
 
 import h5py
-from utils import DEV_STR, TEST_STR, TRAIN_STR, InitYAMLObject
+from utils import DEV_STR, SPLITS, TEST_STR, TRAIN_STR, InitYAMLObject
+from watchdog.observers import Observer
+from watchdog.events import PatternMatchingEventHandler
 from yaml import YAMLObject
+
+
+class _LockFileEventHandler(PatternMatchingEventHandler):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.file_deleted = False
+
+    def on_deleted(self, event):
+        self.file_deleted = True
 
 
 class WholeDatasetCache(InitYAMLObject):
@@ -18,11 +30,47 @@ class WholeDatasetCache(InitYAMLObject):
     cache_checked = False
     yaml_tag = "!WholeDatasetCache"
 
-    def __init__(self, train_path, dev_path, test_path, force_read_cache=False):
-        self.train_path = train_path
-        self.dev_path = dev_path
-        self.test_path = test_path
+    def __init__(
+        self,
+        train_path,
+        dev_path,
+        test_path,
+        force_read_cache=False,
+        wait_for_cache=False,
+    ):
+        self.paths = {
+            TRAIN_STR: train_path,
+            DEV_STR: dev_path,
+            TEST_STR: test_path,
+        }
         self.force_read_cache = force_read_cache
+        self.wait_for_cache = wait_for_cache
+        self.waiting_for_cache = False
+
+    @staticmethod
+    def _sanitize_task_name(task_name):
+        return "".join(
+            [c for c in task_name if c.isalpha() or c.isdigit() or c == " "]
+        ).rstrip()
+
+    @staticmethod
+    def _wait_for_lock(lock_path):
+        """
+        Waits for the file at `lock_path` to be released, i.e. deleted.
+        """
+        event_handler = _LockFileEventHandler(patterns=(os.path.basename(lock_path),))
+        observer = Observer()
+        observer.schedule(event_handler, os.path.dirname(lock_path))
+        observer.start()
+        try:
+            while not event_handler.file_deleted:
+                time.sleep(1)
+        finally:
+            observer.stop()
+            observer.join()
+
+    def _get_cache_path(self, split, task_name):
+        return f"{self.paths[split]}.cache.{task_name}.hdf5"
 
     def get_cache_path_and_check(self, split, task_name):
         """Provides the path for cache files, and cache validity
@@ -37,56 +85,31 @@ class WholeDatasetCache(InitYAMLObject):
              being written to. Does not solve race conditions; use
              cache with caution.)
         """
-        task_name = "".join(
-            [c for c in task_name if c.isalpha() or c.isdigit() or c == " "]
-        ).rstrip()
-        if split == TRAIN_STR:
-            read_cache, write_cache = self.check_cache(self.train_path, task_name)
-            print(
-                "For task {}, split {}, we are reading:{}, writing:{} the cache".format(
-                    task_name, split, read_cache, write_cache
-                )
-            )
-            return (
-                self.train_path + ".cache." + task_name + ".hdf5",
-                read_cache,
-                write_cache,
-            )
-        elif split == DEV_STR:
-            read_cache, write_cache = self.check_cache(self.dev_path, task_name)
-            print(
-                "For task {}, split {}, we are reading:{}, writing:{} the cache".format(
-                    task_name, split, read_cache, write_cache
-                )
-            )
-            return (
-                self.dev_path + ".cache." + task_name + ".hdf5",
-                read_cache,
-                write_cache,
-            )
-        elif split == TEST_STR:
-            read_cache, write_cache = self.check_cache(self.test_path, task_name)
-            print(
-                "For task {}, split {}, we are reading:{}, writing:{} the cache".format(
-                    task_name, split, read_cache, write_cache
-                )
-            )
-            return (
-                self.test_path + ".cache." + task_name + ".hdf5",
-                read_cache,
-                write_cache,
-            )
-        else:
+        if split not in SPLITS:
             return ValueError("Unknown split name: {}".format(split))
+        task_name = self._sanitize_task_name(task_name)
+        cache_path = self._get_cache_path(split, task_name)
+        read_cache, write_cache = self.check_cache(split, cache_path)
+        print(
+            "For task {}, split {}, we are reading:{}, writing:{} the cache".format(
+                task_name, split, read_cache, write_cache
+            )
+        )
+        return cache_path, read_cache, write_cache
 
-    def check_cache(self, path, task_name):
-        """Determines whether datasets have changed; erases caches if so; checks cache lock
+    def check_cache(self, split, cache_path):
+        """
+        Determines whether caches can be read from our written to.
 
-        At the path given, a dataset is required to be there; else an error
-        is thrown. Each file at a path of the form ${path}*.cache is erased
-        if it is older than the file at ${path}.
-        Further, if in this process another object has already started to write
-        to this cache, then use of the cache is disabled.
+        The dataset at `path` must exist, or an error is raised.
+
+        If the caches of the dataset for the specified `task_name`
+        in the form `${path}*.cache` are older than the file at `${path}`,
+        then all caches are erased.
+
+        If the cache's lock file is present, then the cache is not available
+        for reading or writing. If `self.wait_for_cache` is True,
+        then the method blocks until the lock file is released.
 
         Arguments:
           path: The full disk path to a dataset
@@ -95,29 +118,33 @@ class WholeDatasetCache(InitYAMLObject):
             read_cache True if cache should be used for reading, False otherwise
             write_cache True if cache should be written to, False otherwise
         """
-        task_name = "".join(
-            [c for c in task_name if c.isalpha() or c.isdigit() or c == " "]
-        ).rstrip()
-        dataset_time = os.path.getmtime(path)
+        dataset_path = self.paths[split]
+        dataset_time = os.path.getmtime(dataset_path)
         if self.force_read_cache:
+            # Force trying to read from the cache
             print("Forcing trying to read cache, even if not there")
-            return True, False  # Force trying to read from the cache
-        for cache_path in glob.glob(path + "*cache." + task_name + ".*hdf5"):
-            lock_path = cache_path + ".lock"
-            cache_time = os.path.getmtime(cache_path)
-            if cache_time < dataset_time:
-                os.remove(cache_path)
-                logging.info("Cache erased at: {}".format(cache_path))
-                if os.path.exists(lock_path):
-                    os.remove(lock_path)
-                return False, True  # Cache older than data; erased and written to
+            return True, False
+        if not os.path.exists(cache_path):
+            # No cache exists; write one
+            return False, True
+
+        lock_path = cache_path + ".lock"
+        cache_time = os.path.getmtime(cache_path)
+        if cache_time < dataset_time:
+            # Cache is older than data, so erase cache
+            os.remove(cache_path)
+            logging.info("Cache erased at: {}".format(cache_path))
             if os.path.exists(lock_path):
-                return (
-                    False,
-                    False,
-                )  # Cache locked, being written to by object in this process
-            return True, False  # Cache is valid; read from it
-        return False, True  # No cache exists; write one
+                os.remove(lock_path)
+            return False, True
+        if os.path.exists(lock_path):
+            # Cache locked, being written to by this or another process
+            if self.wait_for_cache:
+                print(f"Waiting for {lock_path} to be released...")
+                self._wait_for_lock(lock_path)
+                return self.check_cache(split, cache_path)
+            return False, False
+        return True, False  # Cache is valid; read from it
 
     def release_locks(self):
         """Removes lock files from caches"""
